Refactoring some networking code to make sure that have the proper filenames. Added simple test for NetworkAddress
