+++ src/networking/network_address.cpp
@@ -11,7 +11,7 @@
 //===----------------------------------------------------------------------===//
 
 
-#include "networking/tcp_address.h"
+#include "networking/network_address.h"
 
 #include "common/cast.h"
 #include "common/logger.h"
